fontmaker/requirements.txt
freetype-py==2.5.1
pillow==11.1.0
